TFC-DEISI2053-App4SHM2.0/app4shm-server/requirements/base.txt
Django==4.0
dj-database-url==0.5.0
python-decouple==3.4
pytz==2021.1
djangorestframework==3.13.1
mysqlclient==2.2.4
whitenoise==5.2.0
numpy==1.22.1
scipy==1.7.3
markdown==3.3.6
django-extensions==3.1.5
django-import-export==2.8.0
sentry-sdk==1.44.1